+++ src/main/java/ssu/eatssu/domain/auth/dto/request/AppleLoginRequest.java
@@ -1,4 +1,4 @@
-package ssu.eatssu.domain.auth.dto;
+package ssu.eatssu.domain.auth.dto.request;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import ssu.eatssu.global.log.annotation.LogMask;
