+++ src/main/java/ssu/eatssu/domain/auth/service/OAuthService.java
@@ -6,12 +6,17 @@ import org.springframework.security.config.annotation.authentication.builders.Au
 import org.springframework.security.core.Authentication;
 import org.springframework.stereotype.Service;
 import org.springframework.transaction.annotation.Transactional;
-import ssu.eatssu.domain.auth.dto.*;
+import ssu.eatssu.domain.auth.dto.OAuthInfo;
+import ssu.eatssu.domain.auth.dto.request.AppleLoginRequest;
+import ssu.eatssu.domain.auth.dto.request.AppleLoginRequestV2;
+import ssu.eatssu.domain.auth.dto.request.KakaoLoginRequest;
+import ssu.eatssu.domain.auth.dto.request.KakaoLoginRequestV2;
+import ssu.eatssu.domain.auth.dto.request.ValidRequest;
 import ssu.eatssu.domain.auth.entity.AppleAuthenticator;
 import ssu.eatssu.domain.auth.entity.OAuthProvider;
 import ssu.eatssu.domain.auth.security.JwtTokenProvider;
 import ssu.eatssu.domain.auth.util.RandomNicknameUtil;
-import ssu.eatssu.domain.user.dto.Tokens;
+import ssu.eatssu.domain.user.dto.response.Tokens;
 import ssu.eatssu.domain.user.entity.DeviceType;
 import ssu.eatssu.domain.user.entity.User;
 import ssu.eatssu.domain.user.repository.UserRepository;
