+++ src/test/java/ssu/eatssu/domain/auth/service/OAuthServiceTest.java
@@ -14,14 +14,14 @@ import org.springframework.security.authentication.UsernamePasswordAuthenticatio
 import org.springframework.security.core.Authentication;
 import org.springframework.test.context.ActiveProfiles;
 import org.springframework.transaction.annotation.Transactional;
-import ssu.eatssu.domain.auth.dto.AppleLoginRequest;
-import ssu.eatssu.domain.auth.dto.KakaoLoginRequest;
+import ssu.eatssu.domain.auth.dto.request.AppleLoginRequest;
+import ssu.eatssu.domain.auth.dto.request.KakaoLoginRequest;
 import ssu.eatssu.domain.auth.dto.OAuthInfo;
-import ssu.eatssu.domain.auth.dto.ValidRequest;
+import ssu.eatssu.domain.auth.dto.request.ValidRequest;
 import ssu.eatssu.domain.auth.entity.AppleAuthenticator;
 import ssu.eatssu.domain.auth.entity.OAuthProvider;
 import ssu.eatssu.domain.auth.security.CustomUserDetails;
-import ssu.eatssu.domain.user.dto.Tokens;
+import ssu.eatssu.domain.user.dto.response.Tokens;
 import ssu.eatssu.domain.user.entity.Role;
 import ssu.eatssu.domain.user.entity.User;
 import ssu.eatssu.domain.user.entity.UserStatus;
